backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<3.15
mysqlclient>=2.2.0
python-decouple>=3.8
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
celery>=5.3.0
redis>=5.0.0
django-redis>=5.4.0
hiredis>=2.2.0
boto3>=1.34.0
django-storages>=1.14.0
Pillow>=10.0.0
PyPDF2>=3.0.0
stripe>=7.0.0
requests>=2.31.0
gunicorn>=21.2.0
zoomus>=1.1.0
PyJWT>=2.8.0
openai>=1.0.0
hypothesis>=6.92.0
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
psutil>=5.9.0
python-magic>=0.4.27
sentry-sdk[django]>=1.38.0
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0
